api/requirements.txt
bcrypt==4.3.0
fastapi-users[sqlalchemy]==12.1.2
asyncpg
alembic==1.12.0
psycopg2-binary==2.9.9
pydantic-settings==2.8.1
pillow==10.1.0
pydantic-async-validation==0.1.0
requests==2.31.0
typer==0.15.2
vkbottle==4.5.1
ruff==0.13.2
tenacity==9.1.2
uvicorn==0.38.0
